changes org name
