--- platform_channels/android/app/src/main/kotlin/com/example/platform_channels/MainActivity.kt
@@ -1,31 +1,0 @@
-package com.example.platform_channels
-
-import io.flutter.embedding.android.FlutterActivity
-import io.flutter.embedding.engine.FlutterEngine
-import io.flutter.plugin.common.MethodChannel
-
-class MainActivity : FlutterActivity() {
-    override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
-        // Creates a MethodChannel as soon as the FlutterEngine is attached to
-        // the Activity, and registers a MethodCallHandler. The Method.setMethodCallHandler
-        // is responsible to register a MethodCallHandler to handle the incoming calls.
-
-        // The call parameter of MethodCallHandler has information about the incoming call,
-        // like method name, and arguments. The result parameter of MethodCallHandler is
-        // responsible to send the results of the call.
-        MethodChannel(flutterEngine.dartExecutor, "methodChannelDemo")
-                .setMethodCallHandler { call, result ->
-                    val count: Int? = call.argument<Int>("count")
-
-                    if (count == null) {
-                        result.error("INVALID ARGUMENT", "Value of count cannot be null", null)
-                    } else {
-                        when (call.method) {
-                            "increment" -> result.success(count + 1)
-                            "decrement" -> result.success(count - 1)
-                            else -> result.notImplemented()
-                        }
-                    }
-                }
-    }
-}
